refactor: Update validation rules in AddToCartRequest to require integer quantity; streamline Livewire components by replacing EloquentCollection with Collection for better type consistency; enhance ProductPage with error handling for products without variants and improve global search functionality with conditional Scout integration.

+++ app/Http/Requests/Cart/AddToCartRequest.php
@@ -9,7 +9,7 @@ class AddToCartRequest extends BaseRequest
     public function rules(): array
     {
         return [
-            'quantity' => ['required', 'numeric', 'min:1', 'max:10000'],
+            'quantity' => ['required', 'integer', 'min:1', 'max:10000'],
         ];
     }
 }
